Improvements to weapon menu
Weapon Icons added to weapon menu.
Weapon can be selected from menu.
Only can fire one weapon then turn ends.
Aiming needs implementing.
Fixed bug with turn timer coroutine.

+++ Assets/Scripts/WeaponSystem/BaseWeaponSO.cs
@@ -12,7 +12,7 @@ public class BaseWeaponSO : ScriptableObject {
     public GameObject weaponPrefab = null;
 
     [Header("Weapon Stats")]
-    public float baseDamage = 20.0f;
+    public int baseDamage = 20;
     public float weaponSpeed = 25.0f;
     public bool useGravity = true;
     public bool explosive = false;
